[files] Fix CBlurayFile after 0dc35add3dc6426a531c9903b8ed75fe97a8f5f8.

+++ xbmc/filesystem/BlurayFile.cpp
@@ -62,7 +62,7 @@ namespace XFILE
     return m_file.Stat(RemoveProtocol(url), buffer);
   }
 
-  unsigned int CBlurayFile::Read(void* lpBuf, int64_t uiBufSize)
+  ssize_t CBlurayFile::Read(void* lpBuf, size_t uiBufSize)
   {
     return m_file.Read(lpBuf, uiBufSize);
   }
